dynamic
# message auto-generated for no-merge-commit merge:
merge dynamic_master into master

动态采集模式添加使能失败日志

Created-by: msprofDeveloper

Match-id-2f9478279238cff84a8393592394a87248c693db

+++ analysis/csrc/application/summary/op_summary_assembler.cpp
@@ -251,7 +251,7 @@ void OpSummaryAssembler::WriteToFile(const std::string &fileName, const std::set
 {
     auto timeIndex = GetIndexForVec(headers_, TASK_START_TIME);
     for (auto& row : res_) {
-        if (timeIndex < row.size()) {
+        if (timeIndex < static_cast<int>(row.size())) {
             row[timeIndex].append("\t");
         }
     }
